requirements.txt
torch>=2.0
diffusers>=0.27
transformers>=4.38
accelerate>=0.27
Pillow>=9.0
numpy>=1.24
# optional but recommended for lower VRAM usage:
# xformers
